requirements.txt
astroid==2.3.2
CacheControl==0.12.5
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
deprecation==2.0.7
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.10.3
drf-yasg==1.17.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
jsonpickle==0.9.6
lazy-object-proxy==1.4.3
mailjet-rest==1.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
openapi-codec==1.3.2
packaging==19.2
psycopg2==2.8.4
pylint==2.4.3
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
squareup==3.20191023.0
stripe==2.38.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.4
wrapt==1.11.2
